refactor: separando controller, service e mapper por contexto
