+++ src/main/java/br/com/solutis/locadora/controller/rent/InsurancePolicyController.java
@@ -1,11 +1,12 @@
-package br.com.solutis.locadora.controller;
+package br.com.solutis.locadora.controller.rent;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@Tag(name = "DriverController")
+@Tag(name = "InsurancePolicyController")
 @RestController
-@RequestMapping("/drivers")
-public class DriverController {
+@RequestMapping("/insurances")
+public class InsurancePolicyController {
+
 }
