+++ src/main/java/br/com/solutis/locadora/mapper/person/DriverMapper.java
@@ -1,5 +1,6 @@
-package br.com.solutis.locadora.mapper;
+package br.com.solutis.locadora.mapper.person;
 
+import br.com.solutis.locadora.mapper.GenericMapper;
 import br.com.solutis.locadora.model.dto.DriverDto;
 import br.com.solutis.locadora.model.entity.person.Driver;
 import org.mapstruct.Mapper;
@@ -8,7 +9,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 @Mapper(componentModel = "spring", nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS)
-public interface DriverMapper extends GenericMapper<Driver, DriverDto>{
+public interface DriverMapper extends GenericMapper<Driver, DriverDto> {
     @Override
     DriverDto modelToDTO(Driver entity);
