+++ src/main/java/br/com/solutis/locadora/mapper/rent/InsurancePolicyMapper.java
@@ -1,5 +1,6 @@
-package br.com.solutis.locadora.mapper;
+package br.com.solutis.locadora.mapper.rent;
 
+import br.com.solutis.locadora.mapper.GenericMapper;
 import br.com.solutis.locadora.model.dto.InsurancePolicyDto;
 import br.com.solutis.locadora.model.entity.rent.InsurancePolicy;
 import org.mapstruct.Mapper;
@@ -10,7 +11,7 @@ import java.util.List;
 
 @Component
 @Mapper(componentModel = "spring", nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS)
-public interface InsurancePolicyMapper extends GenericMapper<InsurancePolicy, InsurancePolicyDto>{
+public interface InsurancePolicyMapper extends GenericMapper<InsurancePolicy, InsurancePolicyDto> {
     @Override
     InsurancePolicyDto modelToDTO(InsurancePolicy entity);
