+++ src/main/java/br/com/solutis/locadora/service/person/DriverService.java
@@ -1,9 +1,10 @@
-package br.com.solutis.locadora.service;
+package br.com.solutis.locadora.service.person;
 
-import br.com.solutis.locadora.mapper.DriverMapper;
+import br.com.solutis.locadora.mapper.person.DriverMapper;
 import br.com.solutis.locadora.model.dto.DriverDto;
 import br.com.solutis.locadora.model.entity.person.Driver;
 import br.com.solutis.locadora.repository.CrudRepository;
+import br.com.solutis.locadora.service.CrudService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -14,7 +15,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 @Transactional(propagation = Propagation.REQUIRED)
-public class DriverService implements CrudService<DriverDto>{
+public class DriverService implements CrudService<DriverDto> {
     private final CrudRepository<Driver> driverRepository;
     private final DriverMapper driverMapper;
